requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
openai
orjson
loguru
python-dotenv
pydantic
openai-agents
